move determineBaseRepo to root
